Move Scene from editor to engine, Scene refactoring
